==> notebooks/2.data_creation/2.parse_training.Rmd <==
---
title: "Parse training sets and get URLs"
output: html_notebook
---

```{r message=FALSE}
library(glue)
library(tidyverse)
```


```{r}
info_cols <-
  c(
    "Image_Metadata_MasterIndex",
    "ImageNumber",
    "ObjectNumber",
    "Class",
    "Image_ImageQuality_LocalFocusScore_OrigDNA_5"
  )
```

```{r}
s3 <- paws::s3()

read_data <- function(tag) {
  read_csv(
    glue("input/debris_annotations/MyTrainingSet_{tag}.csv"),
    col_types = cols(.default = col_double(),
                     Class = col_character())
  ) %>%
    select(all_of(info_cols), matches("UserAnnotation_"))
  
}

MyTrainingSet_4channel <- read_data("4channel")
MyTrainingSet_5channel <- read_data("5channel")
MyTrainingSet_6channel <- read_data("6channel")
```

```{r}
get_map <- function(sqlite_file, table_name) {
  con <-
    DBI::dbConnect(RSQLite::SQLite(),
                   glue("input/debris_annotations/{sqlite_file}"))
  qc <- tbl(con, table_name)
  qc %>% select(
    Image_Metadata_MasterIndex,
    Image_Metadata_Plate,
    Image_Metadata_Well,
    Image_Metadata_Site,
    matches("URL|Path"),
    Image_ImageQuality_LocalFocusScore_OrigDNA_5
  ) %>% collect()
}

MyTrainingSet_4channel_map <- 
  get_map("DavidPearlQCFixed_only_nd2_AGP_DNA_ER_Mito.sqlite",
          "DavidPearlQCFixed")
MyTrainingSet_5channel_map <- 
  get_map("DavidPearlQCFixed_exclude_nd2_AGP_DNA_ER_Mito_RNA.sqlite",
          "DavidPearlQCFixed")
MyTrainingSet_6channel_map <- 
  get_map("DavidPearlQCFixed_6channel_exclude_nd2_Actin_DNA_ER_Golgi_Mito_RNA.sqlite",
          "DavidPearlQCFixed_6channel")
```


```{r}
check_map <-
  function(df)
    all(
      df$Image_ImageQuality_LocalFocusScore_OrigDNA_5.x == df$Image_ImageQuality_LocalFocusScore_OrigDNA_5.y
    )

MyTrainingSet_4channel_urls <-
  inner_join(
    MyTrainingSet_4channel,
    MyTrainingSet_4channel_map,
    by = c("Image_Metadata_MasterIndex")
  )

check_map(MyTrainingSet_4channel_urls)

MyTrainingSet_5channel_urls <-
  inner_join(
    MyTrainingSet_5channel,
    MyTrainingSet_5channel_map,
    by = c("Image_Metadata_MasterIndex")
  )

check_map(MyTrainingSet_5channel_urls)

MyTrainingSet_6channel_urls <-
  inner_join(
    MyTrainingSet_6channel,
    MyTrainingSet_6channel_map,
    by = c("Image_Metadata_MasterIndex")
  )

check_map(MyTrainingSet_6channel_urls)
```


```{r}
MyTrainingSet_4channel_urls %>% write_csv("input/debris_annotations/MyTrainingSet_4channel_urls.csv")
MyTrainingSet_5channel_urls %>% write_csv("input/debris_annotations/MyTrainingSet_5channel_urls.csv")
MyTrainingSet_6channel_urls %>% write_csv("input/debris_annotations/MyTrainingSet_6channel_urls.csv")
```
