==> notebooks/2.data_creation/6.inspect-train-test.Rmd <==
---
title: "Inspect training-test splits"
output: html_notebook
---

```{r}
library(glue)
library(tidyverse)
```

```{r}
arpit_counts_summary <-
  read_csv("output/arpit_counts_summary.csv", col_types = cols())
```

```{r}
arpit_qc_annotations_counts <-
  arpit_counts_summary %>%
  mutate(count = ifelse(category == "Good", count_cpa, count_user)) %>%
  select(category, count) %>%
  mutate(Partner = "Broad") %>%
  pivot_wider(names_from = "category", values_from = "count") 

arpit_qc_annotations_counts <-
  arpit_qc_annotations_counts %>%
  mutate(Debris = 208)

arpit_qc_annotations_counts
```


```{r}
partner_annotations_with_metadata <- 
  read_csv("input/partner_annotations_with_metadata.csv",
           col_types = cols(.default = col_character()))

partner_qc_annotations_counts <- 
  partner_annotations_with_metadata %>%
  group_by(Partner, Plate, Class) %>%
  count() %>%
  pivot_wider(names_from = "Class", values_from = "n") %>%
  mutate(across(everything(), ~ replace_na(., 0))) %>%
  rename(Image_Metadata_Plate = Plate)
```


```{r}
all_qc_annotations_counts  <-
  bind_rows(
    arpit_qc_annotations_counts,
    partner_qc_annotations_counts
  )
```


```{r}
all_qc_annotations_counts %>%
  group_by(Partner) %>%
  summarise(across(c(Debris, Empty, Blurry, Good), sum)) 
```

```{r}
all_qc_annotations_counts %>%
  summarise(across(c(Debris, Empty, Blurry, Good), sum)) 
```
